Move fault handling related in ActorCell to ActorCell.FaultHandling.cs
This is to prepare for upcoming refactorings, in order to minimize merge conflicts.
No actual code changes, just moved functions to ActorCell.FaultHandling.cs in the same order as the appear in actor\dungeon\FaultHandling.scala
