+++ src/core/Akka/Akka.csproj
@@ -278,6 +278,7 @@
   <ItemGroup>
     <EmbeddedResource Include="Configuration\Pigeon.conf" />
     <Compile Include="Actor\ActorCell.ReceiveTimeout.cs" />
+    <Compile Include="Actor\ActorCell.FaultHandling.cs" />
     <None Include="Akka.nuspec" />
     <None Include="packages.config" />
     <None Include="Util\MatchHandler\README.md" />
